+++ scripts/fiber/tests-passing-except-dev.txt
@@ -134,6 +134,12 @@ src/renderers/dom/shared/__tests__/ReactDOMComponent-test.js
 * should warn about class (ssr)
 * should suggest property name if available (ssr)
 
+src/renderers/dom/shared/__tests__/ReactDOMServerIntegration-test.js
+* renders a blank div with client render on top of bad server markup
+* renders a div with inline styles with client render on top of bad server markup
+* renders a self-closing tag with client render on top of bad server markup
+* renders a self-closing tag as a child with client render on top of bad server markup
+
 src/renderers/dom/shared/__tests__/ReactMount-test.js
 * should warn if mounting into dirty rendered markup
 * should account for escaping on a checksum mismatch
@@ -144,7 +150,3 @@ src/renderers/dom/shared/__tests__/ReactMountDestruction-test.js
 
 src/renderers/dom/shared/__tests__/ReactServerRendering-test.js
 * should have the correct mounting behavior
-* should render a blank div with client render on top of bad server markup
-* should render a div with inline styles with client render on top of bad server markup
-* should render a self-closing tag with client render on top of bad server markup
-* should render a self-closing tag as a child with client render on top of bad server markup
